Prim
Attached is the main driving script and testing script for Prim's Algorithm implemented using an Adjacency List for the graph and a minHeap for the tree. The implementation was tested using two input text files, city-pairs, and test1. A report listing the results and conclusions for empirical and asymptotic analysis of both implementations will be listed later.